feature(variables): moved createGroqBuilder back into index, since the import order was important
