+++ packages/groq-builder/src/commands/filter.test.ts
@@ -3,7 +3,7 @@ import { SanitySchema, SchemaConfig } from "../tests/schemas/nextjs-sanity-fe";
 import { InferResultType } from "../types/public-types";
 import { executeBuilder } from "../tests/mocks/executeQuery";
 import { mock } from "../tests/mocks/nextjs-sanity-fe-mocks";
-import { createGroqBuilder } from "../createGroqBuilder";
+import { createGroqBuilder } from "../index";
 
 const q = createGroqBuilder<SchemaConfig>();
 const qVariants = q.star.filterByType("variant");
