+++ packages/groq-builder/src/types/public-types.ts
@@ -2,7 +2,7 @@ import type { ZodType } from "zod";
 import type { ResultItem } from "./result-types";
 import type { Simplify } from "./utils";
 import type { ExtractProjectionResult } from "../commands/projection-types";
-import { IGroqBuilder } from "../groq-builder";
+import type { QueryConfig } from "./schema-types";
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
@@ -58,6 +58,34 @@ export type ParserFunctionMaybe<
   TOutput = any
 > = null | ParserFunction<TInput, TOutput>;
 
+// Not used at runtime, just used to infer types:
+export declare const GroqBuilderResultType: unique symbol;
+export declare const GroqBuilderConfigType: unique symbol;
+
+/**
+ * IGroqBuilder is the bare minimum GroqBuilder, used to prevent circular references
+ * @internal
+ */
+export type IGroqBuilder<
+  TResult = unknown,
+  TQueryConfig extends QueryConfig = QueryConfig
+> = {
+  /**
+   * Used to infer the Result types of a GroqBuilder.
+   * This symbol is not used at runtime.
+   * @internal
+   */
+  readonly [GroqBuilderResultType]: TResult;
+  /**
+   * Used to infer the TQueryConfig types of a GroqBuilder.
+   * This symbol is not used at runtime
+   * @internal
+   */
+  readonly [GroqBuilderConfigType]: TQueryConfig;
+  query: string;
+  parser: ParserFunction | null;
+  parse: ParserFunction;
+};
 /**
  * Extracts the Result type from a GroqBuilder query
  */
